# src/routes/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Pick_IRefreshToken.Exclude_keyofIRefreshToken.refreshToken__": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"ip": {
						"type": "string"
					},
					"userAgent": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"_id",
					"ip",
					"date"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IRefreshToken.refreshToken_": {
				"$ref": "#/components/schemas/Pick_IRefreshToken.Exclude_keyofIRefreshToken.refreshToken__",
				"description": "Construct a type with the properties of T except for those in type K."
			}
		},
		"securitySchemes": {
			"access_token": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "backdraft",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/me/logins": {
			"get": {
				"operationId": "Logins",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Omit_IRefreshToken.refreshToken_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"access_token": []
					}
				],
				"parameters": []
			}
		},
		"/me/revoke-login": {
			"delete": {
				"operationId": "RevokeLogin",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"access_token": []
					}
				],
				"parameters": []
			}
		},
		"/me/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"refreshToken": {
													"type": "string"
												}
											},
											"required": [
												"refreshToken"
											],
											"type": "object"
										},
										{}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"access_token": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "user-agent",
						"required": false,
						"schema": {
							"default": "",
							"type": "string"
						}
					}
				]
			}
		},
		"/me/change-email": {
			"post": {
				"operationId": "ChangeEmail",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"access_token": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "host",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "host",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/verify-email": {
			"post": {
				"operationId": "VerifyEmail",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"refreshToken": {
											"type": "string"
										},
										"token": {
											"type": "string"
										}
									},
									"required": [
										"refreshToken",
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "user-agent",
						"required": false,
						"schema": {
							"default": "",
							"type": "string"
						}
					}
				]
			}
		},
		"/token": {
			"post": {
				"operationId": "Token",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										}
									},
									"required": [
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "user-agent",
						"required": false,
						"schema": {
							"default": "",
							"type": "string"
						}
					}
				]
			}
		},
		"/request-password-reset": {
			"post": {
				"operationId": "RequestPasswordReset",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "host",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}